+++ src/common/constants.ts
@@ -106,8 +106,8 @@ export const statusMap: Record<Status, string> = {
 };
 
 // #region to be removed
-// TODO: should be taken from @map-colonies/types
-export interface CallBack<T> {
+// TODO: should be taken from @map-colonies/raster-shared
+export interface CallbackResponse<T> {
   jobId: string;
   taskId: string;
   jobType: string;
@@ -119,6 +119,6 @@ export interface CallBack<T> {
   progress: number;
   message?: string;
   error?: string;
-  params: T;
+  data: T;
 }
 // #endregion to be removed