+++ src/graphql/resolvers/external-services.resolver.ts
@@ -1,9 +1,10 @@
-import { RecordType } from '@map-colonies/mc-model-types';
 import { IConfig } from 'config';
-import { Logger } from 'pino';
 import { container } from 'tsyringe';
 import { Query, Resolver } from 'type-graphql';
+import { Logger } from '@map-colonies/js-logger';
+import { RecordType } from '@map-colonies/mc-model-types';
 import { Services } from '../../common/constants';
+import { extractErrorMessage } from '../../utils';
 import { ExternalService, ServiceType } from '../external-services';
 
 @Resolver((of) => ExternalService)
@@ -21,12 +22,16 @@ export class ExternalServicesResolver {
 
   @Query((type) => [ExternalService])
   public getExternalServices(): ExternalService[] {
-    return this.generateExternalServices();
+    this.logger.info(`[ExternalServices][getExternalServices]`);
+    try {
+      return this.generateExternalServices();
+    } catch (err) {
+      this.logger.error(`[ExternalServices][getExternalServices][ERROR] ${extractErrorMessage(err)}`);
+      throw err;
+    }
   }
 
   private generateExternalServices(): ExternalService[] {
-    this.logger.info(`[ExternalServicesResolver][generateExternalServices] generating external services`);
-
     const extServices = this.externalServices;
     const extServicesArr: ExternalService[] = [];